+++ apps/react/src/components/Header.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { NavLink, useNavigate } from 'react-router-dom';
 import useAuth from '../hooks/useAuth';
-import { apiUrl } from '../lib/jackson';
 
 const Header = () => {
-  const { signOut, user } = useAuth();
+  const { signOut, user, authStatus } = useAuth();
   const navigate = useNavigate();
 
+  const loaded = authStatus === 'LOADED';
   return (
     <header>
       <nav className='border-gray-200 px-4 py-4 shadow'>
@@ -17,36 +17,40 @@ const Header = () => {
                 Home
               </NavLink>
             </li>
-            <li>
-              <NavLink to='/login' className='font-normal text-gray-900'>
-                Login
-              </NavLink>
-            </li>
-            <li>
-              <NavLink to='/profile' className='font-normal text-gray-900'>
-                Profile
-              </NavLink>
-            </li>
-            {user && (
-              <li className='ml-auto'>
-                <button
-                  type='button'
-                  onClick={() =>
-                    signOut(async () => {
-                      // logout from server
-                      const response = await fetch(`${apiUrl}/api/logout`, {
-                        method: 'GET',
-                        credentials: 'include',
-                      });
-                      if (response.status === 401) {
-                        navigate('/login');
-                      }
-                    })
-                  }
-                  className='font-normal text-gray-900'>
-                  Logout
-                </button>
+            {loaded && !user && (
+              <li>
+                <NavLink to='/login' className='font-normal text-gray-900'>
+                  Login
+                </NavLink>
               </li>
+            )}
+            {loaded && user && (
+              <>
+                <li>
+                  <NavLink to='/profile' className='font-normal text-gray-900'>
+                    Profile
+                  </NavLink>
+                </li>
+                <li className='ml-auto'>
+                  <button
+                    type='button'
+                    onClick={() =>
+                      signOut(async () => {
+                        // logout from server
+                        const response = await fetch(`${process.env.REACT_APP_API_URL}/api/logout`, {
+                          method: 'DELETE',
+                          credentials: 'include',
+                        });
+                        if (response.status === 401) {
+                          navigate('/login');
+                        }
+                      })
+                    }
+                    className='font-normal text-gray-900'>
+                    Logout
+                  </button>
+                </li>
+              </>
             )}
           </ul>
         </div>
