+++ apps/react/src/lib/devLogger.ts
@@ -1,0 +1,5 @@
+export default function devLogger(...messages: any[]) {
+  if (process.env.NODE_ENV === 'development') {
+    console.log(...messages);
+  }
+}
