+++ apps/react/src/pages/Profile.tsx
@@ -1,27 +1,8 @@
 import React from 'react';
 import useAuth from '../hooks/useAuth';
 
-// type User = {
-//   id: string;
-//   email: string;
-//   firstName: string;
-//   lastName: string;
-// };
-
 const Profile = () => {
   const { user } = useAuth();
-
-  if (!user) {
-    return (
-      <div className='mx-auto mt-10 max-w-7xl'>
-        <div className='max-w-[50%]'>
-          <div className='rounded border border-gray-200 bg-white py-5 px-5'>
-            <p>Login to access the content.</p>
-          </div>
-        </div>
-      </div>
-    );
-  }
 
   return (
     <div className='mx-auto mt-10 max-w-7xl'>
